requirements.txt
datamodel-code-generator==0.11.12
graphql-core==3.1.6
pydantic==1.8.2